Add files via upload
as per last meeting I added two additional classes Dashboard and Backend and I created the methods and connected the whole classes with appropriate files , I have an issue with the compiler on this pc so i had to upload it without compiling (expect errors). let me know if there is anything i could modify.